requirements.txt
anyconfig==0.10.1
anyio==3.4.0
appdirs==1.4.4
argon2-cffi==21.1.0
arrow==1.2.1
attrs==21.2.0
Babel==2.9.1
backcall==0.2.0
binaryornot==0.4.4
black==21.5b1
bleach==4.1.0
cachetools==4.2.4
certifi==2021.10.8
cffi==1.15.0
chardet==4.0.0
charset-normalizer==2.0.8
click==7.1.2
cookiecutter==1.7.3
coverage==6.2
debugpy==1.5.1
decorator==5.1.0
defusedxml==0.7.1
dynaconf==3.1.5
entrypoints==0.3
flake8==3.9.2
fsspec==2021.11.1
gitdb==4.0.9
GitPython==3.1.24
idna==2.10
importlib-resources==5.4.0
iniconfig==1.1.1
ipykernel==6.6.0
ipython==7.30.1
ipython-genutils==0.2.0
ipywidgets==7.6.5
isort==5.10.1
jedi==0.18.1
Jinja2==3.0.3
jinja2-time==0.2.0
jmespath==0.10.0
joblib==1.1.0
json5==0.9.6
jsonschema==4.2.1
jupyter==1.0.0
jupyter-client==6.1.12
jupyter-console==6.4.0
jupyter-core==4.9.1
jupyter-server==1.12.1
jupyterlab==3.2.4
jupyterlab-pygments==0.1.2
jupyterlab-server==2.8.2
jupyterlab-widgets==1.0.2
kedro==0.17.5
kedro-telemetry==0.1.3
MarkupSafe==2.0.1
matplotlib-inline==0.1.3
mccabe==0.6.1
mistune==0.8.4
mypy-extensions==0.4.3
nbclassic==0.3.4
nbclient==0.5.9
nbconvert==6.3.0
nbformat==5.1.3
nbstripout==0.5.0
nest-asyncio==1.5.4
notebook==6.4.6
numpy==1.21.4
packaging==21.3
pandas==1.3.4
pandocfilters==1.5.0
parso==0.8.3
pathspec==0.9.0
patsy==0.5.2
pexpect==4.8.0
pickleshare==0.7.5
pip-tools==5.5.0
pluggy==0.13.1
poyo==0.5.0
prometheus-client==0.12.0
prompt-toolkit==3.0.23
ptyprocess==0.7.0
py==1.11.0
pycodestyle==2.7.0
pycparser==2.21
pyflakes==2.3.1
Pygments==2.10.0
pyparsing==3.0.6
pyrsistent==0.18.0
pytest==6.2.5
pytest-cov==2.12.1
pytest-mock==1.13.0
python-dateutil==2.8.2
python-json-logger==2.0.2
python-slugify==5.0.2
pytz==2021.3
PyYAML==5.4.1
pyzmq==22.3.0
qtconsole==5.2.1
QtPy==1.11.3
regex==2021.11.10
requests==2.25.1
rope==0.19.0
scikit-learn==1.0.1
scipy==1.7.3
Send2Trash==1.8.0
six==1.16.0
smmap==5.0.0
sniffio==1.2.0
statsmodels==0.13.1
terminado==0.12.1
testpath==0.5.0
text-unidecode==1.3
threadpoolctl==3.0.0
toml==0.10.2
toposort==1.7
tornado==6.1
traitlets==5.1.1
typing_extensions==4.0.1
urllib3==1.26.7
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.2.1
widgetsnbextension==3.5.2
zipp==3.6.0
